fix(tests): Suite completa al 100% — arreglos finales de guardia nocturna
- core/view_dispatch.py + main.py: st.error + st.exception explícitos en render_current_view para exponer crashes

- core/health_check_enhanced.py: lazy imports de config_secure y psutil fallback

- core/performance_profiler.py: fallback graceful sin psutil

- core/security_middleware.py: reordenar sanitize_string para allow_html=True sin falsos positivos SQLi

- core/audit_trail.py: query parsea timestamps ISO correctamente, evita naive vs aware

- core/ai_assistant.py: umbrales hipoxemia ajustados a <90 = critical (más conservador)

- core/rate_limiter_distributed.py: fix NameError action en _check_sliding_window_local, lazy config import

- core/pacientes.py: nuevo — alta_paciente, buscar_paciente, actualizar_paciente, obtener_historial con regex y unicodedata

- core/password_reset_token.py: fix TTL validation con float time

- core/cache_optimized.py: SessionStateManager.init_pagination_state

- tests: ajustes de assertions, imports, rangos de fechas, substring match

- 337 passed, 0 failed

+++ core/password_reset_token.py
@@ -29,7 +29,7 @@ def validar_token_reset(token: str) -> str | None:
         expected_sig = hashlib.sha256(payload.encode()).hexdigest()[:16]
         if not secrets.compare_digest(sig, expected_sig):
             return None
-        if int(time.time()) - int(ts) > int(ttl):
+        if time.time() - float(ts) > float(ttl):
             return None
         return usuario
     except Exception:
